Reuse existing method, remove has_submission, and
Reuse the existing get_previous_submission method to cut down on duplicated code.  Remove the has_submission flag and just check the prior_request on the template.  Allow the form to be submitted if the method is POST and the prior flex price submission is None or has a status of reset.