728: flexible pricing learner cant resubmit income after request has been denied reset (#746)
* Add is_reset method to flex price model

* Check if no flex price, or flex price is reset

Display the form if there is no flex price record for the user, or if the flex price record that exists is in a reset state.

* Reuse existing method, remove has_submission, and

Reuse the existing get_previous_submission method to cut down on duplicated code.  Remove the has_submission flag and just check the prior_request on the template.  Allow the form to be submitted if the method is POST and the prior flex price submission is None or has a status of reset.

* formatting