884: dashboard design update (#888)
* enrolled in free course warning removed

* hold

* UI updates

Move the | after the program ID to within the Start date check, so it only shows when the start date exists.  Only show the course ID and not the full course ID.  Add formatting (background, padding) to the course upgrade description.

* make product page images optional

in order to allow for the default image to be used.

* Format

* Change migration name

* hold but very bR0k3n

* Welp that was quick

* Remove debug statements

* format

* Add serializer test.  Account for anon user

* Add test for flex price API method.  Remove debug.

* enrolled in free course warning removed

* hold

* UI updates

Move the | after the program ID to within the Start date check, so it only shows when the start date exists.  Only show the course ID and not the full course ID.  Add formatting (background, padding) to the course upgrade description.

* make product page images optional

in order to allow for the default image to be used.

* Format

* Change migration name

* hold but very bR0k3n

* Welp that was quick

* Remove debug statements

* format

* Add serializer test.  Account for anon user

* Add test for flex price API method.  Remove debug.

* Fix test

* format

* Fix thee tests