accessibility improvements for dashboard and drawer (#1504)
Link to the program drawer on the dashboard, program tab, program card should read "Program's courses" through a screen reader instead of "2 courses".  Tab navigation should be trapped to the program drawer when it's displayed.  When opening the program drawer, screen readers should pronounce the program title.  Some readers will pronounce a description which would be "X courses, Y passed" where X and Y are the number of courses and those passed, respectively.  The first element of focus in the program drawer should be the "X" close button in the top left.