Code review comments
Fixed regex to work correctly for F/L Name and email.  Name required me to convert it to a string in order to work with HTML `pattern`.  Email required that I escape any special character in the current email before passing that into the regex.  I moved the email validation out of the form into `validation.js` in order to support reuse.  I added positive and negative tests for each of these fields.