715: ecommerce pressing pay jumps back to dashboard without focus on alert (2) (#1544)
* bootstrap updated, much UI broken

* Fixes user menu

* Fixes spacing in forms

* Align form submit buttons correctly

* Fix profile form spacing

* Fix padding and button formatting

Fixes padding issues on upsell message on the enrolled item cards shown on the dashboard.  Also fixes the close button formatting on the program drawer.

* Fix formatting on flexible price form

* Converts font weight for bootstrap 5

Also remove underline from the Title link in the course cards

* Sets the close button to white for alerts

* Move notification component to main from header

* Formatting

* Change reactstrap version

* Chop down flow headaches

* lint

* Fixes course about page

Fixes enroll button sizing, which was caused by the removal of xs sizing.  Replaced those references to xs with sm.  Also fixes formatting on the hamburger menu.

* format

* Make more "data-bs" updates, popover fix

* Remove duplicate prop

* Mobile anonymous navbar works

* Fix links on authenticated user menu