Updating legal address validation to check state validity only if specified (#1574)
The additional info dialog doesn't ask for a state if you're in the US or Canada, so if your account doesn't have one specified, that fails and it results in annoying the learner. So, this updates to only check the state if one is sent in the payload. (The front end validation on that field for the forms that have Country and State will require the learner to specify it there.)