Trying again with migration stuff
Instead of trying to shoehorn the migration in earlier (which will break if you have an existing database, such as on production), I updated a couple of migrations that were pulling Program out of courses.models instead of using the versioned ones (using app.get_model). Seems to have fixed it.

Also added the fields back to the admin