More updates for the enrollment upsell dialog - WIP (stopping for a bit)
- Buttons now actually show the available course runs, with the real dates in them
- Date selector buttons make choices now
- Added a query pull to grab the user's enrollments - we need to deactivate enrollments if they switch dates on us

The user enrollments API may need to be amended a bit so we can filter based on course. Otherwise we're loading all the enrollments to see if they're enrolled in a particular courserun. (Or we update the deactivate enrollment to take a course run in addition to the enrollment ID, or something.)