First run at OIDC login for MITx Online
- Adding in Keycloak instance stuff (mostly copied from UE/Learn but with some reworked documentation)
- Added ol_open_id_connect backend, from Learn, but also captures the global_id (Keycloak ID)
- Added global_id to the user model
- Started in on situating the user for non-API-driven login requests
- Added an interstitial to the login screen to allow the user to choose email or keycloak

This is not particularly usable because OIDC needs feature-flagging. We don't want users trying to log in via Keycloak right yet. (It's also not done. Users get created, that's about it.)

The existing custom parts of the pipeline check to make sure their backend is the one they expect before doing things, so the strategy is to add in basically the existing Python Social Auth pipeline things that would normally be there, but wrapping them so they only run if we're authenticating via OIDC. This should allow the existing email-based flow to work while also allowing OIDC to function alongside it.