Add a function to ensure b2b products have the requisite b2b enrollment code
We were checking to make sure your discounts applied to the products in the basket, but we should probably also make sure you have the discount for the product if you've got a product attached to a B2B contract in your basket too. Otherwise you can check out with a B2B product and pay "full price" for it.