Rework program filtering when an org_id is specified
The filtering was relying on the request user to be filtered through in a manner that wasn't actually workable. Reworked to use the request option that gets passed into the FilterSet anyway.

Also added some code to default filter out programs that require any courses that have B2B courseruns, and update the programs viewset to allow viewing un-authenticated (courses is public access, so should programs, I think, but not B2B ones).