Handle existing baskets better
MITx Online only allows for a single item in the basket at any time. But, this wasn't clearing the existing basket; it was just grabbing it and throwing new stuff in, which confuses the system. So, explicitly clear it (both of items and discounts) before doing anything.

Also adds some tests for this.