Add an API function to reconcile the edX username
If the user has an edX username, do nothing.
If the user doesn't have one, and has an actual user name, use that.
If the user doesn't have one, and doesn't have a a username (or it's the email), generate a username based on name.
If that last one fails, generate one based on the email.