First try at updating the middleware for this
- Adds an API call to reconcile the contract list against the orgs retrieved from the source of truth
- Adds the SSO ID field to the organization page model
- Adds a middleware to debug the APISIX header data (just dumps it out, will get removed)
- Adds a subclass to the APISIX backend to process orgs

Processing the orgs in the backend isn't the best way to do this, but it will be workable for initial testing and local deployments. The problem is that your org membership won't reflect any changes until you log out. It'll get there eventually but since we use this for authorization, it should be faster than this. SCIM will be the better option for this. (But this is fine for local or for now.)

This needs tests still.