Move import_courserun test to better location, pull a bunch of redundant(ish) tests
A lot of the tests here are sort of redundant because most of the logic moved out into an API call (import_courserun_from_edx) that has its own tests. There might be a few that need to be kept, but they should be part of that test suite again and should be reworked to work with the new logic. (Also, this is throwing a bunch of test errors because the mocks aren't right anymore.)