WIP: move UE core APIs over
The UE APIs are really close to the existing ones but we don't have to refactor them a lot to make them work with Spectacular. This also reorganizes the serializers (to add versioning) and redoes some versioning stuff. Very WIP, probably a lot more changes soon.

Also adds an option to the generate_openapi_spec command to allow it to just generate a spec for a single API version, to make it somewhat easier to deal with its output when actively moving stuff around.