Moved the existing views tests into the proper location, moved a bunch of discount and product view tests over, regenerated spec, adjusted URL routing somewhat
This breaks the OpenAPI spec technically - the PR that merged in the basket APIs was only done a couple days ago though. But things were kinda wonky - the basket items API was separate from the basket API, so fixed that so it's now nested under baskets; some of the names were weird so fixed those as well; and fixed some issues with tests (for some reason we were allowing username-based lookup for basket info?).

I also eliminated hard-coded paths in the v0 views tests; probably got a bit annoyed trying to figure out what the route names were for these APIs initially but now that everything is more fixed up it's more straightforward.

Some of the migrated tests are tests for the checkout APIs, which isn't migrated yet. Leaving them in though because that's probably next on the list anyway.